chat/serializer.py
from rest_framework import serializers
from .models import *


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Messages
        fields = ('to_user','from_user','message')
