requirements.txt
Django==1.11
django-cors-headers==2.1.0
djangorestframework==3.6.4
drfdocs==0.0.11
pkg-resources==0.0.0
pytz==2018.3
